chores: clean code a bit
